Make a local copy of main.m, even though it's just the template one.
Also, use Cocoa.h as the prefix header for the project. Even though we don't use CoreData, the default import for most files is Cocoa.h. It could be worth changing them all to the correct headers, or at least Appkit, and indeed some of them have been changed as such, but in the end it's just compile-time effort. Compile time should be better using Cocoa.h as a precompiled header if one or two files accidentally the whole thing later.

git-svn-id: svn+belkadan://belkadan.com/jr/cocoa/webmailer/release@267 ba59ac2a-6797-437e-b805-98a2da23c0da